requirements.txt
langgraph>=0.2.0
pydantic>=2.7
typer>=0.12
rich>=13.7
supabase>=2.0.0
openai>=1.0.0
anthropic>=0.7.0
pinecone-client>=3.0.0
